[WIP] Continuing register cleanup
Not ready for review (very much a work in progress).

Continue the process of using RegStorage rather than
ints to hold register value in the top layers of codegen.
This CL includes some duplicate versions of standard
utilities (some use RegStorage, some use int), but most
of those should be eliminated by the time this is ready
for review.

I'm again attempting this to be a largely mechanical
CL.  The logic changes needed to start taking advantage
of the new register storage model will mostly show up
in later CLs.

Change-Id: I646de7392bdec94595dd2c6f76e0f1c4331096ff