Use MAP_FIXED in MemMap::MapAnonymous().
Thinking more along the line of CL 87711, I think MapAnonymous()
should guarantee the given requested address or else fail. I think
this is how it's intended because the description of MapAnonymous()
says to pass NULL if the caller does not care about the address. (As
an side note, as the image file and the oat file already depend on
MAP_FIXED, this change isn't adding a new dependency on MAP_FIXED.)

There's a potential bug that if MapAnonymous() does not guarantee the
requested address (and there's a gap between the image/oat files and
the zygote/malloc space), then GC could in theory allocate a large
object space in the gap. This would break the GC notion of the immune
space. This change will prevent this by causing all non-moving spaces
to be (really) adjacent, with no gaps in between, which CL 87711
missed.

Change-Id: Id4adb0e30adbad497334d7e00def4c0c66b15719