Restore the args in the artInvokeInterfaceTrampoline slow path.
Previously, there was an error where we did not restore the args for
the slow path in artInvokeInterfaceTrampoline. The slow path occurred
when you attmepted to do an interface dispatch of a method in another
dex file. Since this resolves a method, it can occasionally cause
thread suspension/GC. If a moving GC occurred at this point it
resulted in this_object being invalid after the trampoline returned
the code pointer. This caused a native crash in one of the money runs
since there was an unbundeled app which did a List.iterator call.

Bug: 12934910
Change-Id: Ib454faad14bef0d7732a6d7f1ca5803472d502a6