[WIP]: AArch64: Jni compiler fixes
This patch fixes some of the issues with the ARM64 assembler and JNI
compiler. Currently we do not enable the JNI compiler for ARM64 since
we do not pass the ExceptionHandling test and when generating oat files
some of the alignment checks fail.

To test this patch add:
--- a/compiler/driver/compiler_driver.cc
+++ b/compiler/driver/compiler_driver.cc
@@ -1883,7 +1883,7 @@ void CompilerDriver::CompileMethod(const
DexFile::CodeItem* code_item, uint32_t
   if ((access_flags & kAccNative) != 0) {
     // Are we interpreting only and have support for generic JNI down
     // calls?
     if (!compiler_options_->IsCompilationEnabled() &&
-        (instruction_set_ == kX86_64 || instruction_set_ == kArm64)) {
+        (instruction_set_ == kX86_64)) {
       // Leaving this empty will trigger the generic JNI version
     } else {
       compiled_method = compiler_->JniCompile(*this, access_flags,
method_idx, dex_file);

and build jni_compiler_test

Change-Id: I51d1c24f2c75d4397a11c54724a8b277ff3b3df8
Signed-off-by: Serban Constantinescu <serban.constantinescu@arm.com>