Fix reason for revert.
This change fixes problem caused by the original commit.  It checks
for kArm as well as kThumb2 as the compiler driver and compilation
unit have different ideas of what the instruction set is.

Also fixes a potential issue with deduplication by inserting the
entrypoint offset into to the BL instruction at assembly time.

Change-Id: Ie3d8e1db7c659a406513e73e97047f99da32e5ba