Use scoped arena allocator in OatWriter.
And define template aliases for containers using the
ScopedArenaAllocatorAdapter and use them everywhere.
Make ScopedArenaAllocatorAdapter equality-comparable
and fix DebugStackIndirectTopRefImpl assignment.

This reduces peak memory usage of N5 boot.oat compilation by
6.5MB or more on AOSP ToT. First, we reuse Arenas from the
ArenaPool which usually has 4MB or more already allocated
when we get to the OatWriter. Then we avoid having the four
OatWriter's deduplication maps in memory at the same time.
The largest is the code deduplication map, so we save all
the memory for the other maps which together allocate more
than 2.5MB.

Bug: 13564922
Change-Id: I689aa8a93d169f63a659dec5040567d7b1343277