Don't call SetFootprintLimit in CreateZygoteSpace.
Since SetFootprintLimit uses the allocators accounting it can set a
word inside of the zygote space mem map. This was causing issues
since it was occasionally clobbering one of the object classes. This
occured since we compact objects into the space which was used by the
allocator's internal accounting during zygote bin packing.

Bug: 14056343
Change-Id: I9fc5a61174fa141c0f44d0ed0ae0cc6d77cb4e0a