[WIP] Update load/store utilities for 64-bit targets
Previously, the load/store utilities used sizes of kLong and
kWord - and somewhat sloppily treated kWord as always 32 bits.  This
CL adds a k32 size to explicitly describe the size, and replaces
kLong with k64 for consistency.  kWord remains, but should only be
used when we are dealing with data with the size of the natural
machine word: 64-bits for 64-bit targets, and 32-bits for 32-bit
targets.

In making these changes, it also became clear that it's time to
deal with compressed references.  I'm inclined to be explicit when
loading/storing a reference by requring the use of new load/store
utilites, as well as adding compress and decompress utilities.

Let's discuss.

Change-Id: I50fcbc8684476abd9527777ee7c152c61ba41c6f