ART: Loop Information, initial implementation
Initial version of the loop information API. The two passes are not yet
added into the PassDriver, nor is the information used by anything.
There is already the skeletal induction variable structure that is not filled
yet.

The two passes do the following:

- Find loop: creates the initial loop infrastructure with information about
    the loops.
- Loop Formation: adds the preheader and exit blocks to each loop and fills
    the rest of the information.

There is a skeletal modifications to mir_graph.h, bit_vector.h, compiler_ir.h
that should be removed once the dependencies have merged first.

They are present so that the whole patch can be compiled by all who wish to
test it.

Change-Id: Idf64f996c91b8522f6593afebf25606688f5ed16
Signed-off-by: Jean Christophe Beyler <jean.christophe.beyler@intel.com>
Signed-off-by: Razvan A Lupusoru <razvan.a.lupusoru@intel.com>
Signed-off-by: Yixin Shou <yixin.shou@intel.com>
Signed-off-by: Chao-ying Fu <chao-ying.fu@intel.com>
Signed-off-by: Udayan Banerji <udayan.banerji@intel.com>