[WIP] Move class initialization into the interpreter during invoke
We used to initialize class during invoke in the intepreter bridges. We now
move that part into the interpreter itself and initialize a class after the
resolution but before the actual invoke.

This fixes a stack frame issue where the caller of a <clinit> is incorrect when
calling from intepreter: it was the callee method instead of being the caller
method.

Bug: 14292360
Change-Id: I194d40af5d7f426f89dd6de5bb0e32a6830432b3