ART: Add control flow graph builder [WIP]
A MIRGraph is too heavyweight to represent the control flow graph. For that
reason, the frontend portion has been factored out into a separate class.
The frontend can now be used to build graph for any method and methods
have been added to MIRGraph to be able to integrate a new CFG.

This capability allows the method inliner to build a control flow graph of an
inline target. Then several transformations can be done before it is integrated
into the current compilation unit's MIRGraph. Doing this way also allows
bailing from inliner in case something goes wrong.

A new method has been added to MIRGraph to merge a ControlFlowGraph named
MergeCFG. This updates all of the corresponding MIRGraph data structures
and also prepares callee blocks with names and offsets that can do not overlap
with anything existing in MIRGraph. Although all data structures are updated,
the newly inserted blocks are unreachable unless they are original method
being inserted. The inliner takes care of actually creating the BB links
from caller to callee and vice versa.

Change-Id: I7ba6d6fae4dc9a314b27d09b4e0282b7f3de110d
Signed-off-by: Razvan A Lupusoru <razvan.a.lupusoru@intel.com>