Add Handle/HandleScope and delete SirtRef.
Delete SirtRef and replaced it with Handle. Handles are value types
which wrap around StackReference*.

Renamed StackIndirectReferenceTable to HandleScope.

Added a scoped handle wrapper which wrap around an Object** or
Handle and restores it in its destructor.

Renamed Handle::get -> Get.

TODO: There is some remaining mentions of sirt in variable names and
comments which can be cleaned up.

Bug: 8473721

Change-Id: Idbfebd4f35af629f0f43931b7c5184b334822c7a