ART: Fix contrived verifier error for missing exception class
When verifying an instruction we might have a soft error that we
will interpret as exception-throwing. There are two failure cases:

First, the instruction might be in a try-block. Then a later code
block in verification will try to merge saved registers into catch
blocks, but the merged line wasn't created and contains bogus data.

Second, we do not erase the flag. Thus all following instructions
will be flagged similarly, which increases the chances of the first
error.

Change-Id: I913915ac3202e06dd6e9c24c9d3a7dd3af990d4a