On 64-bit targets, increase the quick frame size by 4 bytes and always flush 8 bytes of Method*.
On 64-bit targets, we want to store 8 bytes of Method* to the quick
frame from the offset 0, so we increase 4 more bytes in the frame.
Also, we always flush 8 bytes of Method* to the frame to workaround
an issue that Method* internally occupies only one compiler temp
register as 4 bytes.
Note that runtime/stack.h assumes Method* is 8 bytes on 64-bit targets.
Without this patch, StackVisitor::SanityCheckFrame() may get
nonzero upper 4 bytes (that should be zeros), when getting
8 bytes of Method*.
If we change Method* to occupy two compiler temps, we will not need
this patch.

Change-Id: I0450bb80298212580ef7917e08f637bd879ba748
Signed-off-by: Chao-ying Fu <chao-ying.fu@intel.com>