Method inlining across dex files in boot image.
Fix LoadCodeAddress() and LoadMethodAddress() to use the dex
file in addition to the method index to uniquely identify
the literal. Add DCHECKs for the assumption that we don't
need to use the invoke type for this. With that fix in place,
when we have both the direct code and the direct method, we
can safely pass the actual target method id instead of the
method id from the same dex file in the method lowering info.
This was already done for calls from apps into boot image
(and thus there was a bug with a tiny risk of the wrong
literal being used) and now we also do that for calls within
the boot image. The latter allows the inlining pass to inline
many more methods than before in the boot image.

Change-Id: Ic765ce9809b43ef07e7db32b8e3fbc9acb09147f