Fix the problem of not shutting down the profiler in ART properly
This CL fixes the problem of not saving the last batch of profile data to the external file when the runtime exits before the end of the current iteration of profile sampling.

I fixed this problem by shutting down the ART profiler when the runtime exits and allowing the profiler to write the last batch of data even it was signaled to shut down.

This CL increases the precision of profile data and fixes the bug that no data is saved if the duration of profile sampling specified by VM arguments (by default 20 seconds) is longer than the length of a single execution because the profiler was never signaled when the runtime exits.

Change-Id: I7cc4805b7e2a22a990d04a5b9724ad1d931d7152
Signed-off-by: Wei Jin <wejin@google.com>